fix default query and update doc
