+++ tests/model_serving/model_server/utils.py
@@ -1,14 +1,11 @@
-import base64
 import json
 import re
 from contextlib import contextmanager
 from string import Template
-from typing import Any, Dict, Generator, List, Optional
+from typing import Any, Dict, Generator, Optional
 
 from kubernetes.dynamic import DynamicClient
 from ocp_resources.inference_service import InferenceService
-from ocp_resources.pod import Pod
-from kubernetes.dynamic.exceptions import ResourceNotFoundError
 from simple_logger.logger import get_logger
 
 from tests.model_serving.model_server.private_endpoint.utils import (
@@ -128,47 +125,6 @@ def _check_storage_arguments(
 ) -> None:
     if (storage_uri and storage_path) or (not storage_uri and not storage_key) or (storage_key and not storage_path):
         raise InvalidStorageArgument(storage_uri, storage_key, storage_path)
-
-
-def b64_encoded_string(string_to_encode: str) -> str:
-    """Returns openshift compliant base64 encoding of a string
-
-    encodes the input string to bytes-like, encodes the bytes-like to base 64,
-    decodes the b64 to a string and returns it. This is needed for openshift
-    resources expecting b64 encoded values in the yaml.
-
-    Args:
-        string_to_encode: The string to encode in base64
-
-    Returns:
-        A base64 encoded string that is compliant with openshift's yaml format
-    """
-    return base64.b64encode(string_to_encode.encode()).decode()
-
-
-def get_pods_by_isvc_label(client: DynamicClient, isvc: InferenceService) -> List[Pod]:
-    """
-    Args:
-        client (DynamicClient): OCP Client to use.
-        isvc (InferenceService):InferenceService object.
-
-    Returns:
-        list[Pod]: A list of all matching pods
-
-    Raises:
-        ResourceNotFoundError: if no pods are found.
-    """
-    if pods := [
-        pod
-        for pod in Pod.get(
-            dyn_client=client,
-            namespace=isvc.namespace,
-            label_selector=f"{isvc.ApiGroup.SERVING_KSERVE_IO}/inferenceservice={isvc.name}",
-        )
-    ]:
-        return pods
-
-    raise ResourceNotFoundError(f"{isvc.name} has no pods")
 
 
 def verify_inference_response(
