+++ utilities/general.py
@@ -1,6 +1,5 @@
+import base64
 from typing import Dict
-
-from tests.model_serving.model_server.utils import b64_encoded_string
 
 
 def get_s3_secret_dict(
@@ -17,3 +16,19 @@ def get_s3_secret_dict(
         "AWS_S3_ENDPOINT": b64_encoded_string(string_to_encode=aws_s3_endpoint),
         "AWS_DEFAULT_REGION": b64_encoded_string(string_to_encode=aws_s3_region),
     }
+
+
+def b64_encoded_string(string_to_encode: str) -> str:
+    """Returns openshift compliant base64 encoding of a string
+
+    encodes the input string to bytes-like, encodes the bytes-like to base 64,
+    decodes the b64 to a string and returns it. This is needed for openshift
+    resources expecting b64 encoded values in the yaml.
+
+    Args:
+        string_to_encode: The string to encode in base64
+
+    Returns:
+        A base64 encoded string that is compliant with openshift's yaml format
+    """
+    return base64.b64encode(string_to_encode.encode()).decode()
