+++ utilities/infra.py
@@ -14,6 +14,7 @@ from ocp_resources.deployment import Deployment
 from ocp_resources.inference_service import InferenceService
 from ocp_resources.infrastructure import Infrastructure
 from ocp_resources.namespace import Namespace
+from ocp_resources.pod import Pod
 from ocp_resources.project_project_openshift_io import Project
 from ocp_resources.project_request import ProjectRequest
 from ocp_resources.role import Role
@@ -23,8 +24,7 @@ from pyhelper_utils.shell import run_command
 from pytest_testconfig import config as py_config
 from simple_logger.logger import get_logger
 
-from tests.model_serving.model_server.utils import b64_encoded_string
-from utilities.general import get_s3_secret_dict
+from utilities.general import b64_encoded_string, get_s3_secret_dict
 
 
 LOGGER = get_logger(name=__name__)
@@ -256,3 +256,28 @@ def get_services_by_isvc_label(client: DynamicClient, isvc: InferenceService) ->
         return svcs
 
     raise ResourceNotFoundError(f"{isvc.name} has no services")
+
+
+def get_pods_by_isvc_label(client: DynamicClient, isvc: InferenceService) -> List[Pod]:
+    """
+    Args:
+        client (DynamicClient): OCP Client to use.
+        isvc (InferenceService):InferenceService object.
+
+    Returns:
+        list[Pod]: A list of all matching pods
+
+    Raises:
+        ResourceNotFoundError: if no pods are found.
+    """
+    if pods := [
+        pod
+        for pod in Pod.get(
+            dyn_client=client,
+            namespace=isvc.namespace,
+            label_selector=f"{isvc.ApiGroup.SERVING_KSERVE_IO}/inferenceservice={isvc.name}",
+        )
+    ]:
+        return pods
+
+    raise ResourceNotFoundError(f"{isvc.name} has no pods")
